requirements.txt
Django>=2.0,<=2.0.2
djangorestframework
django-cors-headers
psycopg2-binary
numpy
pandas
fix_yahoo_finance
datapackage